feat(compiler): 实现编译器后端核心组件
- 完善BasicBlock::removePredecessor方法，确保双向CFG边的一致性删除
- 添加BitVectorTest中的边界检查，防止负索引导致未定义行为
- 新增Aurora编译器项目CMake构建系统配置
- 实现图数据结构ADT用于控制流图表示
- 添加指令选择器实现将AIR指令转换为机器指令
- 新增机器基本块和机器指令的数据结构定义
- 实现aurorac命令行工具及示例函数编译流程
- 添加代码生成管理器及标准编译passes
- 实现线性扫描寄存器分配算法
- 添加选择DAG用于指令选择过程
- 实现SmallVector动态数组容器支持编译器数据结构

+++ tests/ADT/SmallVectorTest.cpp
@@ -25,8 +25,8 @@ TEST(SmallVectorTest, PushBackBeyondInline) {
     for (int i = 0; i < 10; ++i)
         v.push_back(i);
     EXPECT_EQ(v.size(), 10u);
-    for (int i = 0; i < 10; ++i)
-        EXPECT_EQ(v[i], i);
+    for (size_t i = 0; i < 10; ++i)
+        EXPECT_EQ(v[i], static_cast<int>(i));
 }
 
 TEST(SmallVectorTest, PopBack) {
